Issue #9272: Create parser util class that will host util methods for parser that creates AST
